requirements.txt
firebase-admin
openai
python-dotenv
numpy
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic[email]
PyJWT>=2.8.0
APScheduler>=3.10.4
sendgrid>=6.10.0
